refactor: move TimerIncrement into parts
