+++ src/pages/Challenges/index.tsx
@@ -6,6 +6,7 @@ import clsx from 'clsx';
 
 import MainContainer from 'parts/MainContainer';
 import PageTitle from 'parts/PageTitle';
+import TimerIncrement from 'parts/TimerIncrement';
 import VaultScoresTable from 'containers/VaultScoresTable';
 import StakedRelayerScoresTable from 'containers/StakedRelayerScoresTable';
 import InterlayLink from 'components/InterlayLink';
@@ -24,7 +25,6 @@ import {
   POLKA_BTC_DOC_TREASURE_HUNT_VAULT,
   POLKA_BTC_DOC_TREASURE_HUNT_RELAYER
 } from 'config/links';
-import TimerIncrement from 'common/components/timer-increment';
 import { ReactComponent as NewMarkIcon } from 'assets/img/icons/new-mark.svg';
 import { CHALLENGE_CUT_OFFS } from 'config/challenges';
 import 'pages/dashboard/dashboard-subpage.scss';
