+++ src/pages/dashboard/issue/issue.dashboard.page.tsx
@@ -12,12 +12,12 @@ import { BtcNetworkName, IssueColumns } from '@interlay/polkabtc-stats';
 
 import MainContainer from 'parts/MainContainer';
 import PageTitle from 'parts/PageTitle';
+import TimerIncrement from 'parts/TimerIncrement';
 import DashboardTable, {
   StyledLinkData,
   StatusComponent,
   StatusCategories
 } from 'common/components/dashboard-table/dashboard-table';
-import TimerIncrement from 'common/components/timer-increment';
 import LineChartComponent from '../components/line-chart-component';
 import usePolkabtcStats from 'common/hooks/use-polkabtc-stats';
 import { getAccents } from 'pages/dashboard/dashboard-colors';
